[Refactor] Change type from Long to Integer in `CongestionLevelRepository.java`

+++ src/main/java/boombimapi/domain/congestion/repository/CongestionLevelRepository.java
@@ -3,6 +3,6 @@ package boombimapi.domain.congestion.repository;
 import boombimapi.domain.congestion.entity.CongestionLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CongestionLevelRepository extends JpaRepository<CongestionLevel, Long> {
+public interface CongestionLevelRepository extends JpaRepository<CongestionLevel, Integer> {
 
 }
